refactor: 리뷰반영 - Toast, snackBar Util로 관리 @ham2174 @Evergreen97
