+++ presentation/src/main/java/team/jsv/icec/base/BaseActivity.kt
@@ -1,8 +1,6 @@
 package team.jsv.icec.base
 
-import android.content.Context
 import android.os.Bundle
-import android.widget.Toast
 import androidx.annotation.LayoutRes
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
@@ -26,8 +24,4 @@ abstract class BaseActivity<T : ViewDataBinding>(
         binding.unbind()
         super.onDestroy()
     }
-}
-
-fun Context.showToast(msg: String) {
-    Toast.makeText(this, msg, Toast.LENGTH_SHORT).show()
 }
